requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-pymongo==2.3.0
pymongo==4.6.1
bcrypt==4.1.2
anthropic==0.20.0
python-dotenv==1.0.0